requirements.txt
sqlalchemy
requests
beautifulsoup4
python-dotenv
tqdm
